Application/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg[binary]
python-dotenv
pydantic
flask
requests